Crate stageleft

Modules

internal

runtime_support

Macros

PATH_SEPARATOR

q

Creates a quoted expression for Hydroflow+.

quse_fn

stageleft_crate

stageleft_no_entry_crate

Structs

BorrowBounds

RuntimeData

Represents a piece of data that will be passed into the generated code

Traits

IntoQuotedMut

IntoQuotedOnce

Quoted

QuotedContext

Functions

quote_type

Captures a fully qualified path to a given type, which is useful when the generated code needs to explicitly refer to a type known at staging time.

Attribute Macros

entry

Defines an entrypoint for staged code, which will be available as a proc macro. The entrypoint must be a function that returns impl Quoted<T> for some type T.

runtime

Marks a chunk of code as being runtime-only, which means that no staged code in its crate can depend on it. Code behind this attribute is allowed to use staged entrypoints defined in the same crate.

top_level_mod

A utility for declaring top-level public modules in a Stageleft crate that exports macros.