Function stageleft::quote_type

pub fn quote_type<T>() -> Type

Captures a fully qualified path to a given type, which is useful when the generated code needs to explicitly refer to a type known at staging time.

This API is fairly experimental, and comes with caveats. For example, it cannot handle closure types. In addition, when a user refers to a re-exported type, the original type path may be returned here, which could involve private modules.

Also, users must be careful to ensure that any crates referred in the type are available where it is spliced.