Crate stageleft

Modules

internal

runtime_support

Macros

PATH_SEPARATOR

q

Creates a quoted expression for Hydro.

quse_fn

stageleft_crate

stageleft_no_entry_crate

Structs

BorrowBounds

RuntimeData

Represents a piece of data that will be passed into the generated code

Traits

IntoQuotedMut

IntoQuotedOnce

Quoted

QuotedContext

QuotedWithContext

Functions

add_private_reexport

Adds a private module re-export transformation to the type quoting system.

quote_type

Captures a fully qualified path to a given type, which is useful when the generated code needs to explicitly refer to a type known at staging time.

Attribute Macros

entry

Defines an entrypoint for staged code, which will be available as a proc macro. The entrypoint must be a function that returns impl Quoted<T> for some type T.

top_level_mod

A utility for declaring top-level public modules in a Stageleft crate that exports macros.