Module stageleft::internal

Re-exports

pub use proc_macro2;

pub use proc_macro_crate;

pub use syn;

Macros

quote

The whole point.

Structs

Span

A region of source code, along with macro expansion information.

TokenStream

An abstract stream of tokens, or more concretely a sequence of token trees.

Type Aliases

CaptureVec